requirements.txt
SQLAlchemy>=1.4
flask_sqlalchemy>=2.5
sqlalchemy_utils
psycopg2-binary